Splitting functions to Utility directory
